pyproject.toml
[tool.poetry]
name = "technitium-rest-api-client"
version = "0.1.0-dev0"
description = ""
authors = ["Quien Sabe <qs5779@mail.com>"]
readme = "README.md"
packages = [
    { include = "technitium_rac"}
]

[tool.poetry.scripts]
dnsrac = "technitium_rac.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
loguru = "^0.7.2"
pydantic = { version = "^2.8.0", extras = [ "dotenv", "mypy" ] }
yaml-settings-pydantic = {git = "https://github.com/acederberg/pydantic-settings-yaml.git", rev = "2.3.1"}
rest-api-client-framework = {git = "https://github.com/wtfo-guru/rest-api-client-framework.git", rev = "v0.1.1"}

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
mypy = "^1.9"
wemake-python-styleguide = "^0.19"
flake8-pytest-style = ">=1.6,<3.0"
doc8 = "^1.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest-cov = "^5.0.0"
testfixtures = "^8.3.0"
pytest-docker = "^3.1.1"
httpx = "^0.27.0"
safety = "^3.2.7"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=5.2,<8.0"
sphinx-autodoc-typehints = ">=1.24,<3.0"
m2r2 = "^0.3"
tomli = "^2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
